+++ src/auto_survey/pdf_conversion.py
@@ -19,23 +19,57 @@ def convert_markdown_file_to_pdf(markdown_path: Path) -> bool:
     """
     pdf_path = markdown_path.with_suffix(".pdf")
 
-    # Raise an error if Pandoc is not installed
+    # Raise an error if Pandoc and/or WeasyPrint are not installed
     pandoc_installed = (
         subprocess.run(
             ["pandoc", "--version"], capture_output=True, text=True
         ).returncode
         == 0
     )
-    if not pandoc_installed:
-        logger.error(
-            "We cannot convert the Markdown to PDF because Pandoc is not installed. "
-            "Please install Pandoc and try again. Installation instructions can be "
-            "found at https://pandoc.org/installing.html. When Pandoc is installed, "
-            f"you can convert the Markdown at {markdown_path.as_posix()} to PDF by "
-            f"running `pandoc --from=markdown --to=pdf --output={pdf_path.as_posix()} "
-            f"--pdf-engine=weasyprint {markdown_path.as_posix()}` in your terminal."
-        )
-        return False
+    weasyprint_installed = (
+        subprocess.run(
+            ["weasyprint", "--version"], capture_output=True, text=True
+        ).returncode
+        == 0
+    )
+    command_str = (
+        f"`pandoc --from=markdown --to=pdf --output={pdf_path.as_posix()} "
+        f"--pdf-engine=weasyprint {markdown_path.as_posix()}`"
+    )
+    match (pandoc_installed, weasyprint_installed):
+        case (False, False):
+            logger.error(
+                "We cannot convert the Markdown to PDF because both Pandoc and "
+                "WeasyPrint are not installed. Please install both and try again. "
+                "Pandoc installation instructions can be found at "
+                "https://pandoc.org/installing.html and WeasyPrint installation "
+                "instructions can be found at https://doc.courtbouillon.org"
+                "/weasyprint/stable/first_steps.html#installation. When both are "
+                "installed, you can convert the Markdown at "
+                f"{markdown_path.as_posix()} to PDF by running {command_str} in your "
+                "terminal."
+            )
+            return False
+        case (False, True):
+            logger.error(
+                "We cannot convert the Markdown to PDF because Pandoc is not "
+                "installed. Please install it and try again. Installation "
+                "instructions can be found at https://pandoc.org/installing.html. "
+                "When installed, you can convert the Markdown at "
+                f"{markdown_path.as_posix()} to PDF by running {command_str} in your "
+                "terminal."
+            )
+            return False
+        case (True, False):
+            logger.error(
+                "We cannot convert the Markdown to PDF because WeasyPrint is not "
+                "installed. Please install it and try again. Installation "
+                "instructions can be found at https://doc.courtbouillon.org"
+                "/weasyprint/stable/first_steps.html#installation. When installed, you "
+                f"can convert the Markdown at {markdown_path.as_posix()} to PDF by "
+                f"running {command_str} in your terminal."
+            )
+            return False
 
     # Read the Markdown file
     markdown = markdown_path.read_text(encoding="utf-8")