+++ CHANGELOG.md
@@ -9,14 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-- We now check whether `pandoc` is installed before attempting to convert the generated
-  Markdown survey into PDF, and give an informative error message describing how to
-  install these.
 - We've added a logo now, and hidden many of the logs. These can be shown with the
   `--verbose` flag.
 - Removed URLs from the references in the References section.
 - We now use the WeasyPrint PDF engine when converting the Markdown literature survey to
   PDF rather than PDFLaTeX, as the latter often had issues with Unicode characters.
+- We now check whether `pandoc` and `weasyprint` are installed before attempting to
+  convert the generated Markdown survey into PDF, and give an informative error message
+  describing how to install these.
 
 ### Fixed